BASIC LOCO
Cleaned up some of the mounting. Fast low vault is messed up, needs work. Blendspace possibly? Did a SLIDEEEEEEE to crouch. Need to add a max height check for vaulting. Bugs out when you jump near a wall. Next to work on...a few more variations of the vaulting and mantling. Need to fix half height capsule when crouching